reorganize packages for site service module
