+++ core/site/service/endpoint/src/testFixtures/java/org/example/age/service/avs/verification/internal/test/TestFakeAvsVerificationFactoryModule.java
@@ -1,0 +1,10 @@
+package org.example.age.service.avs.verification.internal.test;
+
+import dagger.Module;
+import org.example.age.module.key.common.test.TestKeyModule;
+import org.example.age.service.avs.verification.internal.FakeAvsVerificationFactory;
+import org.example.age.service.avs.verification.internal.FakeAvsVerificationFactoryModule;
+
+/** Dagger module that binds dependencies for {@link FakeAvsVerificationFactory}. */
+@Module(includes = {FakeAvsVerificationFactoryModule.class, TestKeyModule.class})
+public interface TestFakeAvsVerificationFactoryModule {}
